requirements.txt
# --- Core Framework & Integration ---
langchain==0.1.20
langsmith<0.2.0,>=0.1.69
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-core==0.1.52
langchain-text-splitters==0.0.1

# --- Vector Store & Embeddings ---
pinecone-client==3.2.2
openai==1.30.1

# --- Reranking & Search ---
cohere==5.3.2
rank-bm25==0.2.2

# --- Document Processing ---
pypdf==4.2.0
python-docx==1.1.0
# unstructured 0.16.x+ added official support for Python 3.12
unstructured>=0.16.12
tiktoken==0.7.0

# --- API Framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# --- Monitoring & Utilities ---
prometheus-client==0.20.0
python-json-logger==2.0.7
python-dotenv==1.0.1
tenacity==8.3.0
httpx==0.27.0

# --- Data Processing ---
# Numpy 1.26.4 is the minimum stable version for Python 3.12
numpy==1.26.4
pandas==2.2.2

# --- UI ---
streamlit==1.34.0
